abstract generation to library
